requirements.txt
bokeh
streamlit-ace
streamlit-metrics
plotly
cufflinks
nltk
matplotlib
streamlit-folium
altair
folium